requirements.txt
crewai
litellm  # KRITISCH: Benötigt für OpenRouter-Support!
rich
PyYAML
ruamel.yaml  # Für Kommentar-Erhaltung in YAML-Dateien
playwright
Pillow
duckduckgo-search
langchain_community
fastapi
uvicorn
python-multipart
websockets
python-dotenv

# Testing
pytest>=9.0.0
pytest-asyncio>=1.3.0

# Desktop App Testing (ÄNDERUNG 31.01.2026)
pyautogui>=0.9.54
pytest-qt>=4.2.0  # PyQt/PySide UI-Tests (headless-fähig)

# Validation (für Config Validation)
pydantic>=2.8,<3.0.0

# Rate Limiting
slowapi

# Security (für Memory-Verschlüsselung)
cryptography>=41.0.0

# Version-Vergleich (PEP 440, dependency_checker)
packaging>=24.0
